+++ MapboxDirections/Match/MBMatchOptions.swift
@@ -56,6 +56,14 @@ open class MatchOptions: DirectionsOptions {
     @available(*, deprecated: 0.1, message: "Use Waypoint.separatesLegs instead.")
     @objc open var waypointIndices: IndexSet?
     
+    override var legSeparators: [Waypoint] {
+        if let indices = (self as MatchOptionsDeprecations).waypointIndices {
+            return indices.map { super.waypoints[$0] }
+        } else {
+            return super.legSeparators
+        }
+    }
+    
     @objc public required init?(coder decoder: NSCoder) {
         resamplesTraces = decoder.decodeBool(forKey: "resampleTraces")
         super.init(coder: decoder)
@@ -141,22 +149,17 @@ open class MatchOptions: DirectionsOptions {
         let waypoints = namedWaypoints ?? self.waypoints
         let opts = RouteOptions(matchOptions: self)
 
-        var filteredWaypoints: [Waypoint]?
+        let legSeparators: [Waypoint]
         if let indices = (self as MatchOptionsDeprecations).waypointIndices {
-            filteredWaypoints = []
-            for (i, waypoint) in waypoints.enumerated() {
-                if indices.contains(i) {
-                    filteredWaypoints?.append(waypoint)
-                }
-            }
+            legSeparators = indices.map { waypoints[$0] }
         } else {
             waypoints.first?.separatesLegs = true
             waypoints.last?.separatesLegs = true
-            filteredWaypoints = waypoints.filter { $0.separatesLegs }
+            legSeparators = waypoints.filter { $0.separatesLegs }
         }
 
         let routes = (json["matchings"] as? [JSONDictionary])?.map {
-            Route(json: $0, waypoints: filteredWaypoints ?? waypoints, options: opts)
+            Route(json: $0, waypoints: legSeparators, options: opts)
         }
 
         return (waypoints, routes)